requirements.txt
# Core dependencies  
pydantic>=2.0.0  
asyncio  
asyncpg>=0.28.0  
neo4j>=5.0.0  
chromadb>=0.4.0  
numpy>=1.24.0  
tiktoken>=0.5.0  
  
# LLM integrations  
openai>=1.0.0  
google-generativeai>=0.3.0  
  
# Text processing  
regex>=2023.0.0  
nltk>=3.8.0  

# Clustering and ML  
scikit-learn>=1.3.0  
umap-learn>=0.5.0  
networkx>=3.0  
  
# Data processing  
pandas>=2.0.0  
PyYAML>=6.0  
  
# Web framework (optional for API)  
fastapi>=0.100.0  
uvicorn>=0.23.0  
  
# Testing  
pytest>=7.4.0  
pytest-asyncio>=0.21.0  
  
# Development tools  
black>=23.0.0  
flake8>=6.0.0  
mypy>=1.5.0  
  
# Logging  
loguru>=0.7.0  
  
# Environment management  
python-dotenv>=1.0.0

# LLM API
litellm >=1.0.0